x11-display-info: Remove a warning compiling with Meson
Specifically:

../src/x11-display-info.cpp:120:30: warning: comparison of integer expressions of different signedness: 'size_t' {aka 'long unsigned int'} and 'int' [-Wsign-compare]
120 |         for (size_t i = 0; i < drm_resources->count_connectors; ++i) {
    |                            ~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Acked-by: Kevin Pouget <kpouget@redhat.com>

+++ src/x11-display-info.cpp
@@ -117,7 +117,7 @@ public:
     {
         std::vector<std::string> result;
 
-        for (size_t i = 0; i < drm_resources->count_connectors; ++i) {
+        for (int i = 0; i < drm_resources->count_connectors; ++i) {
             drmModeConnectorPtr conn = drmModeGetConnector(drm_fd, drm_resources->connectors[i]);
 
             if (conn == nullptr) {
